+++ components/Header/index.tsx
@@ -1,12 +1,10 @@
 import { Button, Grid } from '@mui/material';
 import { useContext } from 'react';
-import { useAuthContext } from '../../Context/AuthContext';
 import { signOut, useSession } from 'next-auth/react';
 import Text from '@components/atoms/Text';
 import { title } from 'process';
 import LinkTag from '@components/atoms/Link';
 const Header = () => {
-  const authContext = useContext(useAuthContext);
   const { data: session } = useSession();
 
   const nextAuthLogout = () => {
@@ -21,33 +19,42 @@ const Header = () => {
         alignItems={'center'}
         top={0}
       >
-        <Grid item sm={3} sx={{ justifyContent: 'space-around' }} textAlign={'center'}>
+        <Grid
+          item
+          sm={3}
+          sx={{ justifyContent: 'space-around' }}
+          textAlign={'center'}
+        >
           <Text variant={'h5'} text={'title'} />
         </Grid>
         <Grid container item sm={5} sx={{ justifyContent: 'space-around' }}>
-          <Grid item sm={3} style={{ textAlign: 'center' }}>
-            <LinkTag path={'/invoice'}>
-              <Button> create</Button>
-            </LinkTag>
-          </Grid>
-          <Grid item sm={3}>
-            <LinkTag path={'/mylist'}>
-              <Text
-                variant={'h6'}
-                text={'MyList'}
-                style={{ textAlign: 'center' }}
-              />
-            </LinkTag>
-          </Grid>
-          <Grid item sm={3}>
-            <LinkTag path={'/profile'}>
-              <Text
-                variant={'h6'}
-                text={'Profile'}
-                style={{ textAlign: 'center' }}
-              />
-            </LinkTag>
-          </Grid>
+          {session && (
+            <>
+              <Grid item sm={3} style={{ textAlign: 'center' }}>
+                <LinkTag path={'/invoice'}>
+                  <Button> create</Button>
+                </LinkTag>
+              </Grid>
+              <Grid item sm={3}>
+                <LinkTag path={'/mylist'}>
+                  <Text
+                    variant={'h6'}
+                    text={'MyList'}
+                    style={{ textAlign: 'center' }}
+                  />
+                </LinkTag>
+              </Grid>
+              <Grid item sm={3}>
+                <LinkTag path={'/profile'}>
+                  <Text
+                    variant={'h6'}
+                    text={'Profile'}
+                    style={{ textAlign: 'center' }}
+                  />
+                </LinkTag>
+              </Grid>
+            </>
+          )}
         </Grid>
         <Grid container item sm={4} alignItems={'center'}>
           {session?.user?.name ? (
