--- pages/load.tsx
@@ -1,7 +1,0 @@
-import Text from '@components/atoms/Text';
-
-const Load = () => {
-  return <Text variant="h2" text="sushi" />;
-};
-
-export default Load;
