+++ pages/login/index.tsx
@@ -1,18 +1,16 @@
 import React from 'react';
-import { Box, Button, Grid, Typography } from '@mui/material';
+import { Box, Grid} from '@mui/material';
 import { signIn } from 'next-auth/react';
 import Text from '@components/atoms/Text';
 import Image from 'next/image';
 import Illustrationtop from 'public/Illustrationtop.jpg';
 import GoogleButton from 'react-google-button';
 import { useLocale } from 'helper/useLocale';
-import { useRouter } from 'next/router';
 import LoginText from './text.json';
 import { GetServerSideProps } from 'next';
-import BoxLayout from '@components/atoms/Box';
 const Login = (props: any) => {
   const handleGoggle = () => {
-    signIn('google', { callbackUrl: process.env.NEXT_PUBLIC_BASE_URL });
+    signIn('google', { callbackUrl: `${process.env.NEXT_PUBLIC_BASE_URL}/home` });
   };
   const { t } = useLocale(LoginText);
   return (
